feat(DPLAN-18213): Add customField facet
